Safari-611.1.10/LayoutTests/js/dfg-not-string-expected.txt

PASS test() is 9 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
